+++ Example/Utility/ValueArray/main.cpp
@@ -6,7 +6,7 @@
 
 void PerfTest( const size_t size, const size_t n )
 {
-    std::cout << "Performance Test" << std::endl;
+    std::cout << "Performance Test (" << size << " elements, " << n << " times)" << std::endl;
 
     typedef kvs::ValueArray<float> Array;
     const kvs::Indent indent(4);
@@ -87,137 +87,126 @@ int main()
         v1[0] = 1.0f;
         v1[1] = 2.0f;
         v1[2] = 3.0f;
-        std::cout << "kvs::ValueArray<float> v( 3 );" << std::endl;
-        std::cout << indent << "v = " << v1.format() << std::endl;
+        std::cout << indent << "kvs::ValueArray<float> v( 3 );" << std::endl;
+        std::cout << indent.nextIndent() << v1.format() << std::endl;
 
         float temp2[3] = { 2.0f, 3.0f, 4.0f };
         kvs::ValueArray<float> v2( temp2, 3 );
-        std::cout << "float temp[3] = ..." << std::endl;
-        std::cout << "kvs::ValueArray<float> v( temp, 3 );" << std::endl;
-        std::cout << indent << "v2 = " << v2.format() << std::endl;
+        std::cout << indent << "float temp[3] = {...};" << std::endl;
+        std::cout << indent << "kvs::ValueArray<float> v( temp, 3 );" << std::endl;
+        std::cout << indent.nextIndent() << v2.format() << std::endl;
 
-        std::vector<float> temp3;
-        temp3.push_back( 3.0f );
-        temp3.push_back( 4.0f );
-        temp3.push_back( 5.0f );
+        std::vector<float> temp3 = { 3.0f, 4.0f, 5.0f };
         kvs::ValueArray<float> v3( temp3 );
-        std::cout << "std::vector<float> temp; ..." << std::endl;
-        std::cout << "kvs::ValueArray<float> v( temp );" << std::endl;
-        std::cout << indent << "v = " << v3.format() << std::endl;
+        std::cout << indent << "std::vector<float> temp = {...};" << std::endl;
+        std::cout << indent << "kvs::ValueArray<float> v( temp );" << std::endl;
+        std::cout << indent.nextIndent() << v3.format() << std::endl;
 
         kvs::ValueArray<float> v4 = { 4, 5, 6 };
-        std::cout << "kvs::ValueArray<float> v4 = { 4, 5, 6 };" << std::endl;
-        std::cout << indent << "v = " << v4.format() << std::endl;
+        std::cout << indent << "kvs::ValueArray<float> v = { 4, 5, 6 };" << std::endl;
+        std::cout << indent.nextIndent() << v4.format() << std::endl;
         std::cout << std::endl;
     }
 
     std::cout << "Shallow Copy" << std::endl;
     {
-        kvs::ValueArray<float> a( 3 );
-        a[0] = 1.0f;
-        a[1] = 2.0f;
-        a[2] = 3.0f;
+        kvs::ValueArray<int> a = { 1, 2, 3 };
         std::cout << "a = " << a.format() << std::endl;
 
-        kvs::ValueArray<float> b = a; // shallow copy
-        std::cout << "kvs::ValueArray<float> b = a;" << std::endl;
-        std::cout << indent << "b = " << b << std::endl;
+        auto b = a; // shallow copy
+        std::cout << indent << "b = a;" << std::endl;
+        std::cout << indent.nextIndent() << "a: " << a.format() << std::endl;
+        std::cout << indent.nextIndent() << "b: " << b.format() << std::endl;
 
-        a[0] = 0.0f;
-        std::cout << "a[0] = 0.0f;" << std::endl;
-        std::cout << indent << "a = " << a.format() << std::endl;
-        std::cout << indent << "b = " << b.format() << std::endl;
+        a[0] = 0;
+        std::cout << indent << "a[0] = 0;" << std::endl;
+        std::cout << indent.nextIndent() << "a: " << a.format() << std::endl;
+        std::cout << indent.nextIndent() << "b: " << b.format() << std::endl;
         std::cout << std::endl;
     }
 
     std::cout << "Deep Copy" << std::endl;
     {
-        kvs::ValueArray<float> a( 3 );
-        a[0] = 1.0f;
-        a[1] = 2.0f;
-        a[2] = 3.0f;
+        kvs::ValueArray<int> a = { 1, 2, 3 };
         std::cout << "a = " << a.format() << std::endl;
 
-        kvs::ValueArray<float> b = a.clone(); // deep copy
-        std::cout << "kvs::ValueArray<float> b = a.clone();" << std::endl;
-        std::cout << indent << "b = " << b.format() << std::endl;
+        auto b = a.clone(); // deep copy
+        std::cout << indent << "b = a.clone();" << std::endl;
+        std::cout << indent.nextIndent() << "a: " << a.format() << std::endl;
+        std::cout << indent.nextIndent() << "b: " << b.format() << std::endl;
 
-        a[0] = 0.0f;
-        std::cout << "a[0] = 0.0f;" << std::endl;
-        std::cout << indent << "a = " << a.format() << std::endl;
-        std::cout << indent << "b = " << b.format() << std::endl;
+        a[0] = 0;
+        std::cout << indent << "a[0] = 0;" << std::endl;
+        std::cout << indent.nextIndent() << "a: " << a.format() << std::endl;
+        std::cout << indent.nextIndent() << "b: " << b.format() << std::endl;
         std::cout << std::endl;
     }
 
     std::cout << "Assignment" << std::endl;
     {
-        int temp[6] = { 1, 2, 3, 4, 5, 6 };
-        kvs::ValueArray<int> a( temp, 6 );
+        kvs::ValueArray<int> a = { 1, 2, 3, 4, 5, 6 };
         std::cout << "a = " << a.format() << std::endl;
 
         kvs::ValueArray<int> b;
         b.assign( a.begin() + 2, a.end() - 1 );
-        std::cout << "kvs::ValueArray<int> b;" << std::endl;
-        std::cout << "b.assign( a.begin() + 2, a.end() - 1 );" << std::endl;
-        std::cout << indent << "b = " << b.format() << std::endl;
+        std::cout << indent << "b.assign( a.begin() + 2, a.end() - 1 );" << std::endl;
+        std::cout << indent.nextIndent() << "b: " << b.format() << std::endl;
         std::cout << std::endl;
     }
 
     std::cout << "Random" << std::endl;
     {
-        kvs::ValueArray<int> v = kvs::ValueArray<int>::Random( 5 );
-        std::cout << "kvs::ValueArray<int>::Random(n);" << std::endl;
-        std::cout << indent << "v = " << v.format() << std::endl;
+        const size_t n = 5;
+        auto v = kvs::ValueArray<int>::Random( n );
+        std::cout << indent << "kvs::ValueArray<int>::Random( n ); // n = 5" << std::endl;
+        std::cout << indent.nextIndent() << v.format() << std::endl;
 
         const kvs::UInt32 seed = 1;
-        v = kvs::ValueArray<int>::Random( 5, seed );
-        std::cout << "kvs::ValueArray<int>::Random(n,seed);" << std::endl;
-        std::cout << indent << "v = " << v.format() << std::endl;
+        v = kvs::ValueArray<int>::Random( n, seed );
+        std::cout << indent << "kvs::ValueArray<int>::Random( n, seed ); // seed = 1" << std::endl;
+        std::cout << indent.nextIndent() << v.format() << std::endl;
 
-        const int min = 0;
-        const int max = 9;
-        v = kvs::ValueArray<int>::Random( 5, min, max );
-        std::cout << "kvs::ValueArray<int>::Random(n,min,max);" << std::endl;
-        std::cout << indent << "v = " << v.format() << std::endl;
+        const int min = 0, max = 9;
+        v = kvs::ValueArray<int>::Random( n, min, max );
+        std::cout << indent << "kvs::ValueArray<int>::Random( n, min, max ); // min = 0, max = 9" << std::endl;
+        std::cout << indent.nextIndent() << v.format() << std::endl;
 
-        v = kvs::ValueArray<int>::Random( 5, min, max, seed );
-        std::cout << "kvs::ValueArray<int>::Random(n,min,max,seed);" << std::endl;
-        std::cout << indent << "v = " << v.format() << std::endl;
+        v = kvs::ValueArray<int>::Random( n, min, max, seed );
+        std::cout << indent << "kvs::ValueArray<int>::Random( n, min, max, seed );" << std::endl;
+        std::cout << indent.nextIndent() << v.format() << std::endl;
         std::cout << std::endl;
     }
 
     std::cout << "Linear" << std::endl;
     {
-        kvs::ValueArray<int> v1 = kvs::ValueArray<int>::Linear( 5 );
-        std::cout << "kvs::ValueArray<int> v = kvs::ValueArray<int>::Linear( 5 );" << std::endl;
-        std::cout << indent << "v = " << v1.format() << std::endl;
+        const size_t n = 5;
+        auto v1 = kvs::ValueArray<int>::Linear( n );
+        std::cout << indent << "kvs::ValueArray<int>::Linear( n ); // n = 5" << std::endl;
+        std::cout << indent.nextIndent() << v1.format() << std::endl;
 
-        v1 = kvs::ValueArray<int>::Linear( 5, 10 );
-        std::cout << "kvs::ValueArray<int> v = kvs::ValueArray<int>::Linear( 5, 10 );" << std::endl;
-        std::cout << indent << "v = " << v1.format() << std::endl;
+        v1 = kvs::ValueArray<int>::Linear( n, 10 );
+        std::cout << indent << "kvs::ValueArray<int>::Linear( n, 10 ); // start = 10" << std::endl;
+        std::cout << indent.nextIndent() << v1.format() << std::endl;
 
-        v1 = kvs::ValueArray<int>::Linear( 5, 5, 3 );
-        std::cout << "kvs::ValueArray<int> v = kvs::ValueArray<int>::Linear( 5, 5, 3 );" << std::endl;
-        std::cout << indent << "v = " << v1.format() << std::endl;
+        v1 = kvs::ValueArray<int>::Linear( n, 5, 3 );
+        std::cout << indent << "kvs::ValueArray<int>::Linear( n, 5, 3 ); // start = 5, step = 3" << std::endl;
+        std::cout << indent.nextIndent() << v1.format() << std::endl;
 
-        kvs::ValueArray<float> v2 = kvs::ValueArray<float>::Linear( 5, 5.1f );
-        std::cout << "kvs::ValueArray<float> v = kvs::ValueArray<float>::Linear( 5, 5.1f );" << std::endl;
-        std::cout << indent << "v = " << v2.format() << std::endl;
+        auto v2 = kvs::ValueArray<float>::Linear( n, 5.1f );
+        std::cout << indent << "kvs::ValueArray<float>::Linear( n, 5.1 ); // start = 5.1" << std::endl;
+        std::cout << indent.nextIndent() << v2.format() << std::endl;
 
-        v2 = kvs::ValueArray<float>::Linear( 5, 4.9f, 0.1f );
-        std::cout << "kvs::ValueArray<float> v = kvs::ValueArray<float>::Linear( 5, 2.9f, 0.1f );" << std::endl;
-        std::cout << indent << "v = " << v2.format() << std::endl;
-
-        v2 = kvs::ValueArray<float>::Linear( 5, 0.1f, -4.9f );
-        std::cout << "kvs::ValueArray<float> v = kvs::ValueArray<float>::Linear( 5, 0.1f, -2.3f );" << std::endl;
-        std::cout << indent << "v = " << v2.format() << std::endl;
+        v2 = kvs::ValueArray<float>::Linear( n, 2.9f, 0.1f );
+        std::cout << indent << "kvs::ValueArray<float>::Linear( n, 2.9, 0.1 ); // start = 2.9, step = 0.1" << std::endl;
+        std::cout << indent.nextIndent() << v2.format() << std::endl;
         std::cout << std::endl;
     }
 
     std::cout << "Min/Max" << std::endl;
     {
-        kvs::ValueArray<int> v = kvs::ValueArray<int>::Random( 10, 10, 99 );
+        auto v = kvs::ValueArray<int>::Random( 10, 10, 99 );
         std::cout << "v = " << v.format() << std::endl;
+
         std::cout << indent << "min value: " << v.min() << std::endl;
         std::cout << indent << "max value: " << v.max() << std::endl;
         std::cout << indent << "index of min value: " << v.argmin() << std::endl;
@@ -227,34 +216,60 @@ int main()
 
     std::cout << "Shuffle" << std::endl;
     {
-        int temp[6] = { 1, 2, 3, 4, 5, 6 };
-        kvs::ValueArray<int> v( temp, 6 );
+        auto v = kvs::ValueArray<int>::Linear( 5 );
         std::cout << "v = " << v.format() << std::endl;
 
         v.shuffle();
-        std::cout << "v.shuffle();" << std::endl;
-        std::cout << indent << "v = " << v.format() << std::endl;
+        std::cout << indent << "v.shuffle();" << std::endl;
+        std::cout << indent.nextIndent() << v.format() << std::endl;
         std::cout << std::endl;
     }
 
     std::cout << "Sort" << std::endl;
     {
-        kvs::ValueArray<int> v = kvs::ValueArray<int>::Random( 10, 10, 99 );
+        auto v = kvs::ValueArray<int>::Random( 4, 1, 9 );
         std::cout << "v = " << v.format() << std::endl;
 
-        kvs::ValueArray<size_t> i = v.argsort();
-        std::cout << "kvs::ValueArray<size_t> i = v.argsort();" << std::endl;
-        std::cout << indent << "i = " << i.format() << std::endl;
-        std::cout << "kvs::ValueArray<size_t> i = v.argsort( false );" << std::endl;
-        std::cout << indent << "i = " << v.argsort( false ).format() << std::endl;
+        auto i = v.argsort();
+        std::cout << indent << "i = v.argsort();" << std::endl;
+        std::cout << indent.nextIndent() << "i: " << i.format() << std::endl;
+        std::cout << indent << "i = v.argsort( false );" << std::endl;
+        std::cout << indent.nextIndent() << "i: " << v.argsort( false ).format() << std::endl;
 
         v.sort();
-        std::cout << "v.sort();" << std::endl;
-        std::cout << indent << "v = " << v.format() << std::endl;
+        std::cout << indent << "v.sort();" << std::endl;
+        std::cout << indent.nextIndent() << "v: " << v.format() << std::endl;
 
         v.sort( false );
-        std::cout << "v.sort( false );" << std::endl;
-        std::cout << indent << "v = " << v.format() << std::endl;
+        std::cout << indent << "v.sort( false );" << std::endl;
+        std::cout << indent.nextIndent() << "v: " << v.format() << std::endl;
+        std::cout << std::endl;
+    }
+
+    std::cout << "Slice" << std::endl;
+    {
+        auto v = kvs::ValueArray<int>::Linear( 5 );
+        std::cout << "v = " << v.format() << std::endl;
+
+        auto v1 = v.slice( {2} );
+        std::cout << indent << "v1 = v.slice( {2} );" << std::endl;
+        std::cout << indent.nextIndent() << "v1: " << v1.format() << std::endl;
+
+        auto v2 = v.slice( {-2} );
+        std::cout << indent << "v2 = v.slice( {-2} );" << std::endl;
+        std::cout << indent.nextIndent() << "v2: " << v2.format() << std::endl;
+
+        auto v3 = v.slice( {1,3} );
+        std::cout << indent << "v3 = v.slice( {1,3} );" << std::endl;
+        std::cout << indent.nextIndent() << "v3: " << v3.format() << std::endl;
+
+        auto v4 = v.slice( {0,4,2} );
+        std::cout << indent << "v4 = v.slice( {0,4,2} );" << std::endl;
+        std::cout << indent.nextIndent() << "v4: " << v4.format() << std::endl;
+
+        auto v5 = v[ {0,4,2} ];
+        std::cout << indent << "v5 = v[ {0,4,2} ];" << std::endl;
+        std::cout << indent.nextIndent() << "v5: " << v5.format() << std::endl;
         std::cout << std::endl;
     }