+++ Source/Core/Utility/ValueArray.h
@@ -17,6 +17,7 @@
 #include <kvs/DebugNew>
 #include <kvs/Assert>
 #include <kvs/Value>
+#include <kvs/SliceRange>
 #include <kvs/SharedPointer>
 #include <kvs/Deleter>
 
@@ -219,6 +220,11 @@ public:
         return this->data()[ index ];
     }
 
+    ValueArray operator [] ( const kvs::SliceRange& range ) const
+    {
+        return this->slice( range );
+    }
+
     ValueArray& operator =( const ValueArray& rhs )
     {
         ValueArray temp( rhs );
@@ -329,6 +335,23 @@ public:
         return ValueArray( this->data(), this->size() );
     }
 
+    ValueArray slice( const kvs::SliceRange& range ) const
+    {
+        range.adjust( this->size() ); // make the slice indices positive.
+
+        if ( range.step == 1 )
+        {
+            return ValueArray( this->begin() + range.start, this->begin() + range.stop );
+        }
+
+        ValueArray ret( range.size() );
+        for ( long i = range.start, j = 0; i < range.stop; i += range.step )
+        {
+            ret[j++] = this->at(i);
+        }
+        return ret;
+    }
+
     reference at( const size_t index )
     {
         return ( *this )[ index ];
@@ -337,6 +360,11 @@ public:
     const_reference at( const size_t index ) const
     {
         return ( *this )[ index ];
+    }
+
+    ValueArray at( const kvs::SliceRange& range ) const
+    {
+        return ( *this )[ range ];
     }
 
     value_type* pointer()